+++ app/src/androidTest/java/ralcock/cbf/model/dao/BeersImplTest.java
@@ -59,13 +59,13 @@ public class BeersImplTest {
         Brewery brewery2 = new Brewery("2", "Best Brewery", "");
         fBreweries.create(brewery2);
 
-        fBeer1 = new Beer("1", "A Mild", 1f, "description1", fStyle1, "status1", "cask", brewery);
+        fBeer1 = new Beer("1", "A Mild", 1f, "description1", fStyle1, "status1", "cask", "gluten", brewery);
         fBeers.create(fBeer1);
 
-        fBeer2 = new Beer("2", "A Best Bitter", 2f, "description2", fStyle2, "status2", "cask", brewery);
+        fBeer2 = new Beer("2", "A Best Bitter", 2f, "description2", fStyle2, "status2", "cask", "", brewery);
         fBeers.create(fBeer2);
 
-        fBeer3 = new Beer("3", "A Stout", 3f, "description3", fStyle2, "status3", "cask", brewery2);
+        fBeer3 = new Beer("3", "A Stout", 3f, "description3", fStyle2, "status3", "cask", "sulphites", brewery2);
         fBeers.create(fBeer3);
 
         assertEquals(3, fBeers.getNumberOfBeers());
@@ -81,8 +81,10 @@ public class BeersImplTest {
         }
     }
 
-    private List<Beer> doQuery(SortOrder sortOrder, CharSequence filterText, Set<String> stylesToHide) throws SQLException {
-        return fBeers.allBeersList(sortOrder, filterText, stylesToHide, stylesToHide);
+    private List<Beer> doQuery(final SortOrder sortOrder, final CharSequence filterText, final Set<String> stylesToHide) throws SQLException {
+        Set<String> noAllergensToHide = Collections.emptySet();
+        Set<String> noStatusToHide = Collections.emptySet();
+        return fBeers.allBeersList(sortOrder, filterText, stylesToHide, noAllergensToHide, noStatusToHide);
     }
 
     @Test